+++ lib/karafka.rb
@@ -1,7 +1,9 @@
 %w(
+  English
   rake
   rubygems
   bundler
+  English
   celluloid/current
   waterdrop
   pathname
@@ -12,18 +14,15 @@
   sidekiq
   worker_glass
   envlogic
+  thor
+  fileutils
   active_support/callbacks
   active_support/descendants_tracker
   active_support/core_ext/hash/indifferent_access
   active_support/inflector
   karafka/loader
   karafka/status
-  base64
 ).each { |lib| require lib }
-
-# The Poseidon socket timeout is 10, so we give it a bit more time to shutdown after
-# socket timeout
-Celluloid.shutdown_timeout = 15
 
 # Karafka library
 module Karafka
@@ -62,19 +61,31 @@ module Karafka
       Pathname.new(File.expand_path('../..', __FILE__))
     end
 
-    # @return [String] app root path
+    # @return [String] Karafka app root path (user application path)
     def root
       Pathname.new(File.dirname(ENV['BUNDLE_GEMFILE']))
     end
 
-    # @return [String] path to sinatra core root
+    # @return [String] path to Karafka gem root core
     def core_root
       Pathname.new(File.expand_path('../karafka', __FILE__))
+    end
+
+    # @return [String] path to a default file that contains booting procedure etc
+    # @note By default it is a file called 'app.rb' but it can be specified as you wish if you
+    #   have Karafka that is merged into a Sinatra/Rails app and app.rb is taken.
+    #   It will be used for console/workers/etc
+    # @example Standard only-Karafka case
+    #   Karafka.boot_file #=> '/home/app_path/app.rb'
+    # @example Non standard case
+    #   KARAFKA_BOOT_FILE='/home/app_path/karafka.rb'
+    #   Karafka.boot_file #=> '/home/app_path/karafka.rb'
+    def boot_file
+      Pathname.new(
+        ENV['KARAFKA_BOOT_FILE'] || File.join(Karafka.root, 'app.rb')
+      )
     end
   end
 end
 
 Karafka::Loader.new.load!(Karafka.core_root)
-
-load 'karafka/tasks/karafka.rake'
-load 'karafka/tasks/kafka.rake'