+++ lib/karafka/configurators/celluloid.rb
@@ -5,6 +5,9 @@ module Karafka
       # Sets up a Karafka logger as celluloid logger
       def setup
         ::Celluloid.logger = ::Karafka.logger
+        # The Poseidon socket timeout is 10, so we give it a bit more time to shutdown after
+        # socket timeout
+        ::Celluloid.shutdown_timeout = 15
       end
     end
   end